get methods added

+++ homework2-NET/Models/Department.cs
@@ -2,7 +2,7 @@
 {
     public class Department
     {
-        public long DepartmentId { get; set; }
+        public int DepartmentId { get; set; }
         public string DeptName { get; set; }
         public int CountryId { get; set; }
     }
